refactor: separate DOM and error-message function
