+++ login.html
@@ -73,6 +73,6 @@
         <a href="./signup.html" class="signup_btn">회원가입</a>
       </div>
     </div>
-    <script src="./js/login.js"></script>
+    <script src="./script/login.js"></script>
   </body>
 </html>
