+++ script/dom.js
@@ -1,0 +1,7 @@
+export const formElement = document.querySelector(".form");
+export const submitBtn = document.querySelector(".submit-login_btn");
+
+export const emailInput = document.querySelector("#email");
+export const nicknameInput = document.querySelector("#nickname");
+export const passwordInput = document.querySelector("#password");
+export const passwordConfirmInput = document.querySelector("#password_confirm");
